+++ constants/javaTypes.js
@@ -5,3 +5,4 @@ export const S2FPacketSetSlot = Java.type("net.minecraft.network.play.server.S2F
 export const EntityFireworkRocket = Java.type("net.minecraft.entity.item.EntityFireworkRocket");
 export const GuiInventory = Java.type("net.minecraft.client.gui.inventory.GuiInventory");
 export const GuiChat = Java.type("net.minecraft.client.gui.GuiChat");
+export const EntityJoinWorldEvent = Java.type("net.minecraftforge.event.entity.EntityJoinWorldEvent");